refactor: move lambda resources from static to backend
